Fix the parsing of the timestamps
The timestamps of creation, access and modification were parsed
based on the raw output of the "/bin/stat" command.
This parsing was dependant of the text in the output, like the
"Access" word.
On non-English systems this parsing didn't work and the "touch" command
didn't show the timestamps.

The parameter "--printf='%w\n%x\n%y'" has been added to the "/bin/stat"
command so we only get the timestamps of creation, access and
modification. It simplifies the parsing rules and make the parsing
language-agnostic.