Use the dates instead of timestamps in touch
When the user gets a file using Moonwalk, the program gives it the
command to reset the timestamps of creation, modification and access
to their previous values.

When using the timestamps, the precision is no more than seconds, so
in the stat command the dates end with "000000000" for nanoseconds.
It is a hint that the timestamps have been changed.

To avoid this issue, the raw dates as returned by the stat command
are used now. This way, they can be given to the touch command and
the exact same date is reset.